feat: Fix file path for error file in API class
The code changes fix the file path for the error file in the `Api` class. Previously, the file path was specified as "./errors.json", but it should be just "errors.json" to correctly read the file. This ensures that the error messages can be retrieved and added to the API response object when there is an error.

Note: This commit message follows the established convention of starting with a type (feat) followed by a colon and a brief description of the changes.